requirements.txt
asgiref==3.8.1
Django==4.2
django-ranged-response==0.2.0
django-robots==6.1
django-simple-captcha==0.6.1
djangorestframework==3.15.2
pillow==11.1.0
sqlparse==0.5.3
tzdata==2024.2
